backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bycrpt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
alembic==1.13.1
python-dotenv==1.0.0
pyjwt==2.8.0
openai>=1.0.0
slowapi>=0.1.9